Create Plot 3
Calibration of recession probabilities. - Risk graph
Diagonal line is the probability. If an x shows under the we are Over predicting Risk, If there is an x above the line, we are Under predicting the Risk. The size of the x relates to how many months it frll into the same bin.

This plot is optional and it may confuse more than it helps.